fix: buffer sizing, buffer flushing
When doing periodic check, do not write to files, if there is nothing to write.
Saves IO and when compression is enabled.
Does not waste space and does not increase fragmentation by writing empty gzip frames (members).

When checking whether to flush, check if buffer has space for at least one more message.